extracted tests for coordinate conversion
